move standards to separate module
