+++ python/pkg/metric/distance/__init__.py
@@ -1,2 +1,3 @@
 from metric._impl.distance import *
 from metric._impl.voi import *
+from metric._impl.standards import *
