+++ python/pkg/metric/mapping/koc.py
@@ -1,5 +1,5 @@
 import metric._impl.koc as impl
-from metric._impl.distance import Euclidean
+from metric._impl.standards import Euclidean
 from metric._impl.utils import distribution
 from metric._common import factory_with_default
