+++ python/src/distance/init.cpp
@@ -13,7 +13,6 @@ void export_metric_EMD(py::module& m);
 void export_metric_kohonen(py::module& m);
 void export_metric_sorensen(py::module& m);
 void export_metric_SSIM(py::module& m);
-void export_metric_standards(py::module& m);
 void export_metric_TWED(py::module& m);
 void export_metric_Edit(py::module& m);
 
@@ -22,7 +21,6 @@ PYBIND11_MODULE(distance, m) {
     export_metric_kohonen(m);
     export_metric_sorensen(m);
     export_metric_SSIM(m);
-    export_metric_standards(m);
     export_metric_TWED(m);
     export_metric_Edit(m);
 }
